Created longRun ability
Added the ability to leave the bot on for 8 hours at a time, and also have it loop through the file instead of ending when it reached the end of the file. The length of this time can be adjusted by changing the "maxtime" variable in the typeWordsWithoutConfirmation function. Will possibly make this more user-friendly in the future.